Add files via upload
Project to create a competitive Robot team of 3 Robots in the software "Robocode"(https://robocode.sourceforge.io/) to fight in a 3 v 3 v 3 tournament. Each match was a 100 rounds with 3 bots from the above code, 3 bots from the opposing programmers, and 3 standard built-in bots known as "Walls.bot" . The team of bots coded above won the tournament of 14 teams decisively, with 13 victories and 0 losses.